+++ src/index.js
@@ -34,6 +34,7 @@ import SMBGTooltip from './components/daily/smbgtooltip/SMBGTooltip';
 import Stat from './components/common/stat/Stat';
 import CBGTooltip from './components/daily/cbgtooltip/CBGTooltip';
 import FoodTooltip from './components/daily/foodtooltip/FoodTooltip';
+import PhysicalTooltip from './components/daily/physicaltooltip/PhysicalTooltip';
 
 import reducers from './redux/reducers/';
 
@@ -72,6 +73,7 @@ const components = {
   Stat,
   CBGTooltip,
   FoodTooltip,
+  PhysicalTooltip,
 };
 
 const containers = {